refactor(compare): extract C4-C7 to checks/distributions.py (Phase A 7/N)
Move check_c4_mean_preservation, check_c5_mean_after_conversion,
should_run_c5_conversion_check, check_c6_sd_preservation, and
check_c7_categorical_distribution into the checks subpackage. C11
stays in the main file for the next family commit
(checks/type_consistency.py).

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 1524 → 1274 lines.
